modificado:     ejercicios/ejercicio2/.idea/sonarlint/issuestore/1/5/1546737a492b33c1acb736bbb012164b55c55f3f 	nuevo archivo:  ejercicios/ejercicio2/.idea/sonarlint/issuestore/2/6/26b7e5a02eff9415234e77501d2177c54d3bf2fb 	nuevo archivo:  ejercicios/ejercicio2/.idea/sonarlint/issuestore/2/9/299237dfc1b606878f10b86199700f304d5eda8d 	modificado:     ejercicios/ejercicio2/.idea/sonarlint/issuestore/2/d/2d261d756197065627a481271d64949668acb013 	nuevo archivo:  ejercicios/ejercicio2/.idea/sonarlint/issuestore/5/9/592a55bc9094f49bf6d0732d5f5026e06f3f77d5 	nuevo archivo:  ejercicios/ejercicio2/.idea/sonarlint/issuestore/a/7/a7afb116a0c7f21e61bdb4bdcbecc81187bdfbfb 	modificado:     ejercicios/ejercicio2/.idea/sonarlint/issuestore/c/9/c90280a1358748418677580944dbbda99b9742e5 	modificado:     ejercicios/ejercicio2/.idea/sonarlint/issuestore/index.pb 	modificado:     ejercicios/ejercicio2/src/Binario.java 	modificado:     ejercicios/ejercicio2/src/Hexadecimal.java 	modificado:     ejercicios/ejercicio2/src/TestDigitos.java 	modificado:     ejercicios/ejercicio2/src/TestDigitosFichero.java 	modificado:     teoria/.idea/sonarlint/issuestore/0/4/04fc88d9f23821aeac298c2bdbe399ae837f4219 	modificado:     teoria/.idea/sonarlint/issuestore/f/b/fbafd4a96fd0117bb366e588b6e5052b535adc2e 	modificado:     teoria/.idea/vcs.xml

+++ ejercicios/ejercicio2/src/Hexadecimal.java
@@ -14,7 +14,7 @@ public class Hexadecimal {
         return valor.matches("[0-9abcdefABCDEF]+");
     }
 
-    public int devolverValorDecimal(){
+    public long devolverValorDecimal(){
        /* double suma = 0.0;
         StringBuilder sB = new StringBuilder(valor.toLowerCase()); //para poder acceder al reverse
         sB.reverse();  //StringBuilder es mutable, cambia
@@ -49,7 +49,7 @@ public class Hexadecimal {
             }
         }
         return (int) suma; */
-        return Integer.parseInt(valor, 16);
+        return Long.parseLong(valor, 16);
 
     }
 
@@ -60,9 +60,9 @@ public class Hexadecimal {
                 valor, devolverValorDecimal());
     }
 
-    public static int obtenerSumaDecimalListaHexadecimales(
+    public static long obtenerSumaDecimalListaHexadecimales(
             List<Hexadecimal> lista ) {
-        int suma = 0;
+        long suma = 0;
         for ( Hexadecimal h: lista ) {
             suma += h.devolverValorDecimal();
         }
